services/legal-compliance/requirements-test.txt
pytest==7.3.1
pytest-asyncio==0.21.0
pytest-cov==4.1.0
httpx==0.24.1
requests==2.29.0
PyJWT==2.7.0